fix: fix toolset auth title in marketplace
